refactor(): create package blackjack.model.states for all state representing classes, and delete .gitkeep files
