+++ src/main/kotlin/blackjack/model/Participant.kt
@@ -1,5 +1,7 @@
 package blackjack.model
 
+import blackjack.model.states.State
+
 abstract class Participant {
     abstract val name: String
     abstract var state: State
