+++ src/main/kotlin/blackjack/model/states/BlackJack.kt
@@ -1,9 +1,11 @@
-package blackjack.model
+package blackjack.model.states
+
+import blackjack.model.HandCards
 
 class BlackJack(
     override val hand: HandCards = HandCards(),
     override val rate: Double = 1.5
-) : Stop() {
+) : Stopped() {
     init {
         if (hand.total() != 21 && hand.size() != 2) { throw IllegalStateException() }
     }
