+++ src/main/kotlin/blackjack/model/states/Bust.kt
@@ -1,9 +1,11 @@
-package blackjack.model
+package blackjack.model.states
+
+import blackjack.model.HandCards
 
 class Bust(
     override val hand: HandCards,
     override val rate: Double = -1.0
-) : Stop() {
+) : Stopped() {
     init {
         if (hand.total() <= 21) { throw IllegalStateException() }
     }
