+++ src/main/kotlin/blackjack/model/states/Stay.kt
@@ -1,9 +1,11 @@
-package blackjack.model
+package blackjack.model.states
+
+import blackjack.model.HandCards
 
 class Stay(
     override val hand: HandCards,
     override val rate: Double = 1.0
-) : Stop() {
+) : Stopped() {
     init {
         if (hand.size() < 2) { throw IllegalStateException() }
     }
