+++ src/test/kotlin/blackjack/BustTest.kt
@@ -1,6 +1,6 @@
 package blackjack
 
-import blackjack.model.Bust
+import blackjack.model.states.Bust
 import blackjack.model.HandCards
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
